requirements.txt
biopython>=1.79
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
torch-geometric>=2.0.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
